requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langgraph==0.4.1
faiss-cpu==1.9.0
langchain-google-genai
python-dotenv==1.1.0
google-generativeai